update of referred rs categories while updating r/s candidate

+++ offering_catalog/src/main/java/it/nextworks/tmf_offering_catalog/services/ResourceCandidateService.java
@@ -39,7 +39,36 @@ public class ResourceCandidateService {
     @Autowired
     private ResourceCategoryService resourceCategoryService;
 
-    public ResourceCandidate create(ResourceCandidateCreate resourceCandidateCreate) throws NotExistingEntityException, NullIdentifierException {
+    private void updateResourceCategoryCreate(ResourceCandidate rc)
+            throws NullIdentifierException, NotExistingEntityException {
+
+        List<ResourceCategoryRef> resourceCategoryRefs = rc.getCategory();
+        if(resourceCategoryRefs != null) {
+            for(ResourceCategoryRef resourceCategoryRef : resourceCategoryRefs) {
+
+                String rcrId = resourceCategoryRef.getId();
+                if(rcrId == null)
+                    throw new NullIdentifierException("Referenced Resource Category with null identified not allowed.");
+
+                ResourceCategory rCategory = resourceCategoryService.get(rcrId);
+
+                log.info("Updating Resource Category " + rcrId + ".");
+
+                rCategory.getResourceCandidate().add(new ResourceCandidateRef()
+                        .href(rc.getHref())
+                        .id(rc.getId())
+                        .lastUpdate(OffsetDateTime.ofInstant(Instant.now(), ZoneId.of("UTC")).toString())
+                        .name(rc.getName()));
+
+                resourceCategoryService.save(rCategory);
+
+                log.info("Resource Category " + rcrId + " updated.");
+            }
+        }
+    }
+
+    public ResourceCandidate create(ResourceCandidateCreate resourceCandidateCreate)
+            throws NotExistingEntityException, NullIdentifierException {
 
         log.info("Received request to create a Resource Candidate.");
 
@@ -62,29 +91,7 @@ public class ResourceCandidateService {
         if(lastUpdate != null)
             resourceCandidate.setLastUpdate(lastUpdate.toString());
 
-        List<ResourceCategoryRef> resourceCategoryRefs = resourceCandidate.getCategory();
-        if(resourceCategoryRefs != null) {
-            for(ResourceCategoryRef resourceCategoryRef : resourceCategoryRefs) {
-
-                String rcrId = resourceCategoryRef.getId();
-                if(rcrId == null)
-                    throw new NullIdentifierException("Referenced Resource Category with null identified not allowed.");
-
-                ResourceCategory rCategory = resourceCategoryService.get(rcrId);
-
-                log.info("Updating Resource Category " + rcrId + ".");
-
-                rCategory.getResourceCandidate().add(new ResourceCandidateRef()
-                        .href(resourceCandidate.getHref())
-                        .id(id)
-                        .lastUpdate(OffsetDateTime.ofInstant(Instant.now(), ZoneId.of("UTC")).toString())
-                        .name(resourceCandidate.getName()));
-
-                resourceCategoryService.save(rCategory);
-
-                log.info("Resource Category " + rcrId + " updated.");
-            }
-        }
+        updateResourceCategoryCreate(resourceCandidate);
 
         resourceCandidateRepository.save(resourceCandidate);
 
@@ -93,15 +100,8 @@ public class ResourceCandidateService {
         return resourceCandidate;
     }
 
-    public void delete(String id) throws NotExistingEntityException, NullIdentifierException {
-
-        log.info("Received request to delete Resource Candidate with id " + id + ".");
-
-        Optional<ResourceCandidate> toDelete = resourceCandidateRepository.findByResourceCandidateId(id);
-        if(!toDelete.isPresent())
-            throw new NotExistingEntityException("Resource Candidate with id " + id + " not found in DB.");
-
-        ResourceCandidate rc = toDelete.get();
+    private void updateResourceCategoryDelete(ResourceCandidate rc)
+            throws NullIdentifierException, NotExistingEntityException {
 
         List<ResourceCategoryRef> resourceCategoryRefs = rc.getCategory();
         if(resourceCategoryRefs != null) {
@@ -124,6 +124,19 @@ public class ResourceCandidateService {
                 log.info("Resource Category " + rcrId + " updated.");
             }
         }
+    }
+
+    public void delete(String id) throws NotExistingEntityException, NullIdentifierException {
+
+        log.info("Received request to delete Resource Candidate with id " + id + ".");
+
+        Optional<ResourceCandidate> toDelete = resourceCandidateRepository.findByResourceCandidateId(id);
+        if(!toDelete.isPresent())
+            throw new NotExistingEntityException("Resource Candidate with id " + id + " not found in DB.");
+
+        ResourceCandidate rc = toDelete.get();
+
+        updateResourceCategoryDelete(rc);
 
         resourceCandidateRepository.delete(rc);
 
@@ -147,7 +160,7 @@ public class ResourceCandidateService {
     }
 
     public ResourceCandidate patch(String id, ResourceCandidateUpdate resourceCandidateUpdate, String lastUpdate)
-            throws NotExistingEntityException {
+            throws NotExistingEntityException, NullIdentifierException {
 
         log.info("Received request to patch Resource Candidate with id " + id + ".");
 
@@ -162,14 +175,19 @@ public class ResourceCandidateService {
         resourceCandidate.setType(resourceCandidateUpdate.getType());
 
         final List<ResourceCategoryRef> category = resourceCandidateUpdate.getCategory();
-        if(resourceCandidate.getCategory() == null)
+        if(resourceCandidate.getCategory() == null) {
             resourceCandidate.setCategory(category);
-        else if(category != null) {
+            updateResourceCategoryCreate(resourceCandidate);
+        } else if(category != null) {
+            updateResourceCategoryDelete(resourceCandidate);
             resourceCandidate.getCategory().clear();
+
             resourceCandidate.getCategory().addAll(category);
-        }
-        else
+            updateResourceCategoryCreate(resourceCandidate);
+        } else {
+            updateResourceCategoryDelete(resourceCandidate);
             resourceCandidate.getCategory().clear();
+        }
 
         resourceCandidate.setDescription(resourceCandidateUpdate.getDescription());
         resourceCandidate.setLastUpdate(resourceCandidateUpdate.getLastUpdate().toString());
